more spelling adjustments
