+++ custom_components/thermal_comfort/utils.py
@@ -1,0 +1,14 @@
+"""Utils for Thermal comfort."""
+import math
+
+from homeassistant.const import STATE_UNAVAILABLE, STATE_UNKNOWN
+
+
+def _is_valid_state(state) -> bool:
+    if state is not None:
+        if state.state not in (STATE_UNKNOWN, STATE_UNAVAILABLE):
+            try:
+                return not math.isnan(float(state.state))
+            except ValueError:
+                pass
+    return False
